+++ bindings/mnt/pyfiction/include/pyfiction/pybind11_mkdoc_docstrings.hpp
@@ -10648,6 +10648,13 @@ static const char *__doc_fiction_detail_technology_mapping_impl_stats = R"doc(Te
 
 static const char *__doc_fiction_detail_technology_mapping_impl_technology_mapping_impl = R"doc()doc";
 
+static const char *__doc_fiction_detail_technology_mapping_impl_validate_required_gates =
+R"doc(Validate that the technology library contains the required gates for
+the base network type.
+
+Throws:
+    missing_required_gates_exception if required gates are missing.)doc";
+
 static const char *__doc_fiction_detail_to_hex =
 R"doc(Utility function to transform a Cartesian tile into a hexagonal one.
 
@@ -16967,6 +16974,12 @@ Returns:
     Iterator to the minimum energy charge distribution found in the
     input range, or `last` if the range is empty.)doc";
 
+static const char *__doc_fiction_missing_required_gates_exception =
+R"doc(Exception thrown when a technology mapping library does not contain
+the required gates for the base network type.)doc";
+
+static const char *__doc_fiction_missing_required_gates_exception_missing_required_gates_exception = R"doc()doc";
+
 static const char *__doc_fiction_missing_sidb_position_exception =
 R"doc(Exception thrown when a missing SiDB position is encountered in the
 parsed file.)doc";
@@ -21920,7 +21933,7 @@ static const char *__doc_fiction_technology_dot_drawer_node_label_callback = R"d
 static const char *__doc_fiction_technology_mapping =
 R"doc(Performs technology mapping on the given network. Technology mapping
 is the process of replacing the gates in a network with gates from a
-given technology library. This function utilizes `mockturtle::map` to
+given technology library. This function utilizes `mockturtle::emap` to
 perform the technology mapping. This function is a wrapper around that
 interface to provide a more convenient usage.
 
@@ -21959,7 +21972,7 @@ static const char *__doc_fiction_technology_mapping_params_ge2 = R"doc(2-input g
 
 static const char *__doc_fiction_technology_mapping_params_gt2 = R"doc(2-input greater-than gate.)doc";
 
-static const char *__doc_fiction_technology_mapping_params_inv = R"doc()doc";
+static const char *__doc_fiction_technology_mapping_params_inv = R"doc(1-input NOT gate (inverter).)doc";
 
 static const char *__doc_fiction_technology_mapping_params_le2 = R"doc(2-input less-or-equal gate.)doc";
 
@@ -21969,7 +21982,7 @@ static const char *__doc_fiction_technology_mapping_params_maj3 = R"doc(3-input 
 
 static const char *__doc_fiction_technology_mapping_params_mapper_params = R"doc(mockturtle's mapper parameters.)doc";
 
-static const char *__doc_fiction_technology_mapping_params_mux = R"doc(3-input MUX gate.)doc";
+static const char *__doc_fiction_technology_mapping_params_mux = R"doc(3-input MUX gate (ITE).)doc";
 
 static const char *__doc_fiction_technology_mapping_params_nand2 = R"doc(2-input NAND gate.)doc";